refactor: move getUserDisplayLabel to users/utils to remove cross-feature dep
